load_tests/requirements.txt
locust>=2.20.0
websockets>=12.0
psycopg2-binary>=2.9.0
aiohttp>=3.9.0
tabulate>=0.9.0
